refactor: 把漫画目录的信息独立成一个vue组件
