+++ src/utils.ts
@@ -1,0 +1,8 @@
+import {commands} from "./bindings.ts";
+
+export async function showPathInFileManager(path: string | undefined) {
+    if (path === undefined) {
+        return;
+    }
+    await commands.showPathInFileManager(path);
+}
